Fix more warnings

+++ DemoGame/src/SpriteRendererSystem.cpp
@@ -25,10 +25,10 @@ void SpriteRendererSystem::Render(EntityContainer& entityContainer, SDL_Renderer
 
 		Texture* texture = spriteRenderer.texture;
 		Vec2f screenPosition = cameraComponent.ConvertFromWorldPositionToScreenPosition(transform.GetPosition());
-		destRect.x = screenPosition.X() - (texture->GetDimensions().w / 2);
-		destRect.y = screenPosition.Y() - (texture->GetDimensions().h / 2);
-		destRect.w = texture->GetDimensions().w * transform.GetScale().X();
-		destRect.h = texture->GetDimensions().h * transform.GetScale().Y();
+		destRect.x = static_cast<int>(screenPosition.X() - (texture->GetDimensions().w / 2.f));
+		destRect.y = static_cast<int>(screenPosition.Y() - (texture->GetDimensions().h / 2.f));
+		destRect.w = static_cast<int>(texture->GetDimensions().w * transform.GetScale().X());
+		destRect.h = static_cast<int>(texture->GetDimensions().h * transform.GetScale().Y());
 
 		//SDL_RenderCopy(renderer, texture->GetRaw(), &texture->GetDimensions(), &destRect);
 		SDL_RenderCopyEx(renderer, texture->GetRaw(), &texture->GetDimensions(), &destRect, transform.GetRotationAngle(), nullptr, SDL_RendererFlip::SDL_FLIP_NONE);
